Reuse colcon_ros package identification
