+++ colcon_ros_cargo/task/ament_cargo/build.py
@@ -2,6 +2,7 @@
 
 import os
 from pathlib import Path
+import subprocess
 
 from colcon_cargo.task.cargo import CARGO_EXECUTABLE
 from colcon_cargo.task.cargo.build import CargoBuildTask
@@ -47,6 +48,16 @@ class AmentCargoBuildTask(CargoBuildTask):
             '.cargo/config.toml for subsequent builds with cargo.')
 
     def _prepare(self, env, additional_hooks):
+        ament_build = 'cargo ament-build --help'.split()
+        if subprocess.run(ament_build, capture_output=True).returncode != 0:
+            if _print_ament_cargo_warning_once():
+                logger.error(
+                    '\n\nament_cargo package found but cargo ament-build was '
+                    'not detected.'
+                    '\n\nPlease install it by running:'
+                    '\n $ cargo install cargo-ament-build\n')
+            return 1
+
         args = self.context.args
 
         global package_paths
@@ -92,6 +103,23 @@ class AmentCargoBuildTask(CargoBuildTask):
     # Installation is done by cargo ament-build
     def _install_cmd(self, cargo_args):  # noqa: D102
         pass
+
+
+def _print_ament_cargo_warning_once():
+    global has_printed_ament_cargo_warning
+    try:
+        # The following line will throw an exception if the global variable
+        # has never been initialized
+        has_printed_ament_cargo_warning
+    except NameError:
+        # We want to initialize the global variable to false the first time
+        has_printed_ament_cargo_warning = False
+
+    if not has_printed_ament_cargo_warning:
+        has_printed_ament_cargo_warning = True
+        return True
+
+    return False
 
 
 def write_cargo_config_toml(package_paths):
