+++ setup.cfg
@@ -70,12 +70,10 @@ markers =
     linter
 
 [options.entry_points]
-colcon_core.package_identification =
-    ament_cargo = colcon_ros_cargo.package_identification.ament_cargo:AmentCargoPackageIdentification
 colcon_core.task.build =
-    ament_cargo = colcon_ros_cargo.task.ament_cargo.build:AmentCargoBuildTask
+    ros.ament_cargo = colcon_ros_cargo.task.ament_cargo.build:AmentCargoBuildTask
 colcon_core.task.test =
-    ament_cargo = colcon_ros_cargo.task.ament_cargo.test:AmentCargoTestTask
+    ros.ament_cargo = colcon_ros_cargo.task.ament_cargo.test:AmentCargoTestTask
 
 [flake8]
 import-order-style = google
